Trying to get Milight API to load
Not at all successfull.
Got a bunch of attempted and commented out lines of code trying to load.

Not sure if everything should go in libraries or 3rd party.
Found a few links saying namespaces (and maybe interfaces?) are not
supported by CI b/c it was written to support PHP 4.
So 2 options - I can use those links to use autoload funcitonality
(altered) or manual load to get my stuff in.  This looks pretty tricky.
Or I could make an attempt to rewrite the API into individual class
files with no dependencies.  I might like to try this.  If it works I
could upload to GitHub as specific CodeIgniter API.   Would be a good
challenge.  But definitley doable I think b/c the existing API is pretty
understandable.